+++ src/bit_battle.c
@@ -36,6 +36,8 @@ SOFTWARE.
 ////////////////////////////////////////////////////////////
 
 #include <stdio.h>
+#include <stdlib.h>
+#include <math.h>
 #include <raylib.h>
 
 ////////////////////////////////////////////////////////////
@@ -44,7 +46,84 @@ SOFTWARE.
 
 #include "bittendef.h"
 #include "bit_battle.h"
-void bit_BattleDraw(int* hp)
+
+////////////////////////////////////////////////////////////
+// Just some conversion tools
+////////////////////////////////////////////////////////////
+
+
+// Reverses a string 'str' of length 'len'
+void reverse(char* str, int len)
 {
-    
+    int i = 0, j = len - 1, temp;
+    while (i < j) {
+        temp = str[i];
+        str[i] = str[j];
+        str[j] = temp;
+        i++;
+        j--;
+    }
+}
+ 
+// Converts a given integer x to string str[].
+// d is the number of digits required in the output.
+// If d is more than the number of digits in x,
+// then 0s are added at the beginning.
+int intToStr(int x, char str[], int d)
+{
+    int i = 0;
+    while (x) {
+        str[i++] = (x % 10) + '0';
+        x = x / 10;
+    }
+ 
+    // If number of digits required is more, then
+    // add 0s at the beginning
+    while (i < d)
+        str[i++] = '0';
+ 
+    reverse(str, i);
+    str[i] = '\0';
+    return i;
+}
+ 
+// Converts a floating-point/double number to a string.
+void ftoa(float n, char* res, int afterpoint)
+{
+    // Extract integer part
+    int ipart = (int)n;
+ 
+    // Extract floating part
+    float fpart = n - (float)ipart;
+ 
+    // convert integer part to string
+    int i = intToStr(ipart, res, 0);
+ 
+    // check for display option after point
+    if (afterpoint != 0) {
+        res[i] = '.'; // add dot
+ 
+        // Get the value of fraction part upto given no.
+        // of points after dot. The third parameter
+        // is needed to handle cases like 233.007
+        fpart = fpart * pow(10, afterpoint);
+ 
+        intToStr((int)fpart, res + i + 1, afterpoint);
+    }
+
+}
+
+////////////////////////////////////////////////////////////
+// battle display
+////////////////////////////////////////////////////////////
+
+void bit_BattleDraw(float* playerHPw, char** enemyw, float* enemyHPw, int* screenWidthw, int* screenHeightw)
+{
+    DrawText("Well That was easy", 190, 200, 20, BLACK);
+    DrawText(*enemyw, *screenWidthw/4*3, *screenHeightw/4*3, 10, BLACK);
+    char working[5];
+    ftoa(*playerHPw, working, 4);
+    DrawText(working, *screenWidthw/4, *screenHeightw/4*3, 10, BLACK);
+    ftoa(*playerHPw, working, 4);
+    DrawText(working, *screenWidthw/4*3, *screenHeightw/4*3.1, 10, BLACK);
 }
