+++ src/main.c
@@ -48,73 +48,8 @@ SOFTWARE.
 
 #include "bittendef.h" // defines for the engine
 #include "bit_loadfile.h" // file loading functionality
-// settings.h for when  i go to fix that trainwreck of a file
-//#include "settings.h"
+#include "bit_battle.h"
 
-////////////////////////////////////////////////////////////
-// Just some conversion tools
-////////////////////////////////////////////////////////////
-
-
-// Reverses a string 'str' of length 'len'
-void reverse(char* str, int len)
-{
-    int i = 0, j = len - 1, temp;
-    while (i < j) {
-        temp = str[i];
-        str[i] = str[j];
-        str[j] = temp;
-        i++;
-        j--;
-    }
-}
- 
-// Converts a given integer x to string str[].
-// d is the number of digits required in the output.
-// If d is more than the number of digits in x,
-// then 0s are added at the beginning.
-int intToStr(int x, char str[], int d)
-{
-    int i = 0;
-    while (x) {
-        str[i++] = (x % 10) + '0';
-        x = x / 10;
-    }
- 
-    // If number of digits required is more, then
-    // add 0s at the beginning
-    while (i < d)
-        str[i++] = '0';
- 
-    reverse(str, i);
-    str[i] = '\0';
-    return i;
-}
- 
-// Converts a floating-point/double number to a string.
-void ftoa(float n, char* res, int afterpoint)
-{
-    // Extract integer part
-    int ipart = (int)n;
- 
-    // Extract floating part
-    float fpart = n - (float)ipart;
- 
-    // convert integer part to string
-    int i = intToStr(ipart, res, 0);
- 
-    // check for display option after point
-    if (afterpoint != 0) {
-        res[i] = '.'; // add dot
- 
-        // Get the value of fraction part upto given no.
-        // of points after dot. The third parameter
-        // is needed to handle cases like 233.007
-        fpart = fpart * pow(10, afterpoint);
- 
-        intToStr((int)fpart, res + i + 1, afterpoint);
-    }
-}
 
 ////////////////////////////////////////////////////////////
 // Entrypoint of Engine
@@ -223,17 +158,9 @@ int main(int argc, char **argv){
         BeginDrawing();
             ClearBackground(WHITE);
             if (title) DrawText("bitten's adventure", 190, 200, 20, BLACK);
-            if (!title) DrawTMX(map, x, y, WHITE);
+            if (!title && !battle) DrawTMX(map, x, y, WHITE);
             DrawTextureRec(bitten,bittenRec,bittenPos,WHITE);
-            if (battle){
-                DrawText("Well That was easy", 190, 200, 20, BLACK);
-                DrawText(enemy, screenWidth/4*3, screenHeight/4*3, 10, BLACK);
-                char working[5];
-                ftoa(playerHP, working, 4);
-                DrawText(working, screenWidth/4, screenHeight/4*3, 10, BLACK);
-                ftoa(playerHP, working, 4);
-                DrawText(working, screenWidth/4*3, screenHeight/4*3.1, 10, BLACK);
-            }
+            if (battle) bit_BattleDraw(&playerHP, &enemy, &enemyHP, &screenWidth, &screenHeight);
             DrawFPS(10, 10);
         EndDrawing();
     }
